consolidate ScrollableIterator
