+++ elide-datastore/elide-datastore-hibernate/src/main/java/com/yahoo/elide/core/hibernate/ScrollableIteratorBase.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2021, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.core.hibernate;
+
+import com.google.common.base.Preconditions;
+import com.google.common.collect.Iterators;
+
+import lombok.NonNull;
+
+import java.util.ConcurrentModificationException;
+import java.util.Iterator;
+import java.util.function.Function;
+import java.util.function.Predicate;
+
+/**
+ * Wraps ScrollableResult as Iterator.
+ *
+ * @param <T> type of return object
+ * @param <R> Result object
+ */
+public class ScrollableIteratorBase<T, R> implements Iterable<T>, Iterator<T> {
+    private final R scroll;
+    private final Predicate<R> nextFun;
+    private final Function<R, T> getFun;
+    private boolean inUse = false;
+    private boolean hasNext = false;
+
+    public ScrollableIteratorBase(R scroll, Predicate<R> nextFun, Function<R, T> getFun) {
+        this.scroll = scroll;
+        this.nextFun = nextFun;
+        this.getFun = getFun;
+
+        hasNext = nextFun.test(scroll);
+    }
+
+    @Override
+    public Iterator<T> iterator() {
+        if (inUse) {
+            throw new ConcurrentModificationException();
+        }
+
+        inUse = true;
+        return Iterators.unmodifiableIterator(this);
+    }
+
+    @Override
+    public boolean hasNext() {
+        return hasNext;
+    }
+
+    @Override
+    @NonNull public T next() {
+        @NonNull T row = getFun.apply(scroll);
+        Preconditions.checkNotNull(row);
+        hasNext = nextFun.test(scroll);
+        return row;
+    }
+}
