+++ elide-datastore/elide-datastore-hibernate5/src/main/java/com/yahoo/elide/datastores/hibernate5/ScrollableIterator.java
@@ -5,51 +5,17 @@
  */
 package com.yahoo.elide.datastores.hibernate5;
 
-import com.google.common.base.Preconditions;
-import com.google.common.collect.Iterators;
-import org.hibernate.ScrollableResults;
-import lombok.NonNull;
+import com.yahoo.elide.core.hibernate.ScrollableIteratorBase;
 
-import java.util.ConcurrentModificationException;
-import java.util.Iterator;
+import org.hibernate.ScrollableResults;
 
 /**
  * Wraps ScrollableResult as Iterator.
  *
  * @param <T> type of return object
  */
-public class ScrollableIterator<T> implements Iterable<T>, Iterator<T> {
-    private final ScrollableResults scroll;
-    private boolean inUse = false;
-    private boolean hasNext = false;
-
-    public ScrollableIterator(ScrollableResults scroll) {
-        this.scroll = scroll;
-
-        hasNext = scroll.next();
-    }
-
-    @Override
-    public Iterator<T> iterator() {
-        if (inUse) {
-            throw new ConcurrentModificationException();
-        }
-
-        inUse = true;
-        return Iterators.unmodifiableIterator(this);
-    }
-
-    @Override
-    public boolean hasNext() {
-        return hasNext;
-    }
-
-    @Override
-    @NonNull public T next() {
-        @SuppressWarnings("unchecked")
-        @NonNull T row = (T) scroll.get()[0];
-        Preconditions.checkNotNull(row);
-        hasNext = scroll.next();
-        return row;
+public class ScrollableIterator<T> extends ScrollableIteratorBase<T, ScrollableResults> {
+    public ScrollableIterator(ScrollableResults results) {
+        super(results, ScrollableResults::next, scroll -> (T) scroll.get()[0]);
     }
 }
